Safari-609.1.11.1/LayoutTests/js/function-dot-apply-replace-base-expected.txt

This tests that expressions of the form x = x.apply(...) don't break bytecode generation.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS bar() is 3
PASS successfullyParsed is true

TEST COMPLETE
